# feat: emit diagnostics_channel events upon routing request