--- a/lib/handleRequest.js
+++ b/lib/handleRequest.js
@@ -125,7 +125,7 @@ function preHandlerCallback (err, request, reply) {
 
   const context = request[kRouteContext]
 
-  if (!tcHasSubscribers() || context[kFourOhFourContext] === null) {
+  if (!channels.hasSubscribers || context[kFourOhFourContext] === null) {
     preHandlerCallbackInner(err, request, reply)
   } else {
     const store = {
@@ -180,14 +180,6 @@ function preHandlerCallbackInner (err, request, reply, store) {
   } finally {
     if (store) channels.end.publish(store)
   }
-}
-
-function tcHasSubscribers () {
-  return channels.start.hasSubscribers ||
-    channels.end.hasSubscribers ||
-    channels.asyncStart.hasSubscribers ||
-    channels.asyncEnd.hasSubscribers ||
-    channels.error.hasSubscribers
 }
 
 module.exports = handleRequest