--- a/package.json
+++ b/package.json
@@ -206,7 +206,7 @@
     "@fastify/fast-json-stringify-compiler": "^4.3.0",
     "abstract-logging": "^2.0.1",
     "avvio": "^8.3.0",
-    "dc-polyfill": "^0.1.5",
+    "dc-polyfill": "^0.1.6",
     "fast-json-stringify": "^5.14.1",
     "find-my-way": "^8.1.0",
     "light-my-request": "^5.13.0",