@mcollina 2024-01-04
> Is there a way to get the payload sent via reply.send()?

inside `reply.send()` or when executing the onSend hook.