@Uzlopak 2024-01-04
What about the diagnostics_channel plugin?