@tlhunter 2024-01-04
> What about the diagnostics_channel plugin?

@Uzlopak I think you're asking about the removal of the `diagnostics_channel` npm package?

If so, the `dc-polyfill` package makes the `diagnostics_channel` package obsolete. The way `diagnostics_channel` worked is that it was a snapshot of the internal `diagnostics_channel` module that was made into an npm package of the same name. On ancient Node.js versions where the internal `diagnostics_channel` module was missing a call to `require('diagnostics_channel')` would pickup the npm package and that same call on newer versions of Node.js simply retrieves the internal module. However there were shortcomings with the `diagnostics_channel` package, like if multiple versions of the `diagnostics_channel` package existed in the a `node_modules/` dependency tree then they wouldn't cross communicate. Also the `diagnostics_channel` package hasn't been maintained and is missing newer features that this PR depends on. `dc-polyfill` alleviates those issues and many more.

If anyone is curious please checkout the [dc-polyfill](https://github.com/DataDog/dc-polyfill) docs for more info. Ultimately `dc-polyfill` is a drop-in replacement for both the `diagnostics_channel` npm package and internal module