@tlhunter 2024-01-05
> inside `reply.send()` or when executing the onSend hook.

@mcollina I'm now grabbing the payload within `Reply` instances and setting it on itself using a new symbol: https://github.com/fastify/fastify/pull/5252/commits/92f77e4b1d183a873bc750ded6710514883b9052

That said I'm not sure if there are expectations on how long `Reply` instances stick around and if there may be memory concerns. Let me know if this approach isn't ideal and I'll try something else.