@mcollina 2024-01-05
>  I wonder if we should integrate then the whole diagnostics-channel stuff from the plugin into core. 

Agreed