@Uzlopak 2024-01-11
@mcollina 

Should this PR also integrate the plugins behaviour? Or should this be a follow up?