@mcollina 2024-01-12
> @mcollina 
> 
> 
> 
> Should this PR also integrate the plugins behaviour? Or should this be a follow up?

Can you clarify?