@tlhunter 2024-01-12
@mcollina I think @Uzlopak is asking if I should copy all of the code from the existing `fastify-diagnostics-channel` package and incorporate it into my PR. I'm of the opinion that it shouldn't be included in this PR. While that package is semi-related in scope, it uses a different pattern, and this PR is just attempting to implement the TracingChannel "spec". Plus I'm going to write some blog posts that link to this PR to share with the community and I'd like to keep the PR as succinct and isolated as possible to avoid confusing the audience.