@tlhunter 2024-01-12
@Qard and I paired on this PR today. He envisions two levels of tracing channel in web frameworks. The outer level is that of routing, and the inner level is that of request handling. Previously I was taking a sort of hybrid approach. This PR has been refactored to implement tracing at the routing level. For that reason code has been pulled out of `reply.js` and moved into `wrapThenable.js`.

The response payload isn't necessarily required for APM tools, and with the way TracingChannel is designed it's a bit kludgy to extract using `res.send()` anyway. While it's easy to extract in some scenarios (returns/resolves) I've modified the PR to not include `.result` in the emitted store objects to reduce confusion.

Adding tracing at the request level is something that can be added in the future. The two will play nicely with each other at that point. Request tracing would require the changes to `res.send()` that I was previously working on. Request level tracing is mostly covered by http tracing which is why it's a lower priority.

Since this is the routing layer tracing I also updated the stores to include information about the routing decision made by Fastify. Specifically the underlying route pattern and HTTP method.

Tracing plugins/middleware would be valuable too but we're currently considering them to be out of scope.