@mcollina 2024-01-13
I think the nomenclature and names do not match what we use in Fastify. I would identify that specific event as “request.handler”, because “rout ing” happens in a completely different place.