@Uzlopak 2024-01-14
The tests are failing and seem to be a valid report, because handleRequest.js was modified by this PR.